+++ addOns/selenium/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Support for selenium scripts which are invoked when browsers are launched.
+
 ### Changed
 - Update minimum ZAP version to 2.9.0.
 - Set Firefox browser.tabs.documentchannel pref to false to fix HUD issue